refactor: move pattern matchers to their respective modules
